Some changes

+++ src/Vect.h
@@ -120,9 +120,9 @@ class Vect
 	{
 		for(int i = 0; i < this->size(); i++)
 			if( i != this->size() - 1 )
-				printf("[%f],", this->data[i]);
+				printf("[%f],", (double)this->data[i]);
 			else 
-				printf("[%f]\n", this->data[i]);
+				printf("[%f]\n", (double)this->data[i]);
 	}
         //double Angle()const {return atan2(Y2, long double Y1);}
 };
